README.md
A/B tests are very commonly performed by data analysts and data scientists. It is important to get some practice working with the difficulties of this kind of experiment due to its frequency of use.
For this project, we will understand the results of an A/B test run by an e-commerce website. The goal is to work through this notebook to help the company understand if they should implement the new page, keep the old page, or perhaps run the experiment longer to make their decision.